release: publish v1.5.0

+++ src/background/services/sidebarMessageRouter.js
@@ -9,9 +9,12 @@ import {
 import { getLlmConfig, saveLlmConfig } from './llmService.js';
 import {
     clearProdMemoCache,
+    clearProdMemoSyncData,
     deleteProdMemoCache,
+    exportProdMemoCache,
     getProdMemoCache,
     importProdMemoCache,
+    runProdMemoAction,
 } from './prodMemoService.js';
 import {
     clearSessionKeeperLogs,
@@ -57,8 +60,11 @@ chrome.runtime.onMessage.addListener((msg, sender, sendResponse) => {
     if (msg.type === 'WQP_SESSION_TOKEN_CAPTURED') {
         return respond(sendResponse, handleCapturedSessionToken(msg.token));
     }
-    if (msg.type === 'WQP_PRODMEMO_GET' || msg.type === 'WQP_PRODMEMO_EXPORT') {
+    if (msg.type === 'WQP_PRODMEMO_GET') {
         return respond(sendResponse, getProdMemoCache());
+    }
+    if (msg.type === 'WQP_PRODMEMO_EXPORT') {
+        return respond(sendResponse, exportProdMemoCache());
     }
     if (msg.type === 'WQP_PRODMEMO_IMPORT') {
         return respond(sendResponse, importProdMemoCache(msg.memoData));
@@ -66,8 +72,14 @@ chrome.runtime.onMessage.addListener((msg, sender, sendResponse) => {
     if (msg.type === 'WQP_PRODMEMO_CLEAR') {
         return respond(sendResponse, clearProdMemoCache());
     }
+    if (msg.type === 'WQP_PRODMEMO_CLEAR_SYNC') {
+        return respond(sendResponse, clearProdMemoSyncData());
+    }
     if (msg.type === 'WQP_PRODMEMO_DELETE') {
         return respond(sendResponse, deleteProdMemoCache(msg.alphaId));
+    }
+    if (msg.type === 'WQP_PRODMEMO_DB') {
+        return respond(sendResponse, runProdMemoAction(msg.action, msg.payload || {}));
     }
     if (msg.type === 'WQP_LLM_CONFIG_GET') {
         return respond(sendResponse, getLlmConfig());
